+++ stroom-index/stroom-index-impl/src/main/java/stroom/index/impl/rest/IndexVolumeResourceImpl.java
@@ -24,7 +24,7 @@ public class IndexVolumeResourceImpl implements IndexVolumeResource {
     }
 
     @Override
-    public Response getById(final Long id) {
+    public Response getById(final int id) {
         final IndexVolume indexVolume = indexVolumeService.getById(id);
         return Response.ok(indexVolume).build();
     }
@@ -38,7 +38,7 @@ public class IndexVolumeResourceImpl implements IndexVolumeResource {
     }
 
     @Override
-    public Response delete(final Long id) {
+    public Response delete(final int id) {
         indexVolumeService.delete(id);
         return Response.noContent().build();
     }
@@ -51,7 +51,7 @@ public class IndexVolumeResourceImpl implements IndexVolumeResource {
     }
 
     @Override
-    public Response addVolumeToGroup(final Long volumeId,
+    public Response addVolumeToGroup(final int volumeId,
                                      final String groupName) {
         indexVolumeService.addVolumeToGroup(volumeId, groupName);
 
@@ -59,7 +59,7 @@ public class IndexVolumeResourceImpl implements IndexVolumeResource {
     }
 
     @Override
-    public Response removeVolumeFromGroup(final Long volumeId,
+    public Response removeVolumeFromGroup(final int volumeId,
                                           final String groupName) {
         indexVolumeService.removeVolumeFromGroup(volumeId, groupName);
