+++ stroom-index/stroom-index-mock/src/main/java/stroom/index/mock/MockIndexVolumeService.java
@@ -30,12 +30,12 @@ public class MockIndexVolumeService implements IndexVolumeService {
     }
 
     @Override
-    public IndexVolume getById(Long id) {
+    public IndexVolume getById(int id) {
         return null;
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(int id) {
 
     }
 
@@ -45,12 +45,12 @@ public class MockIndexVolumeService implements IndexVolumeService {
     }
 
     @Override
-    public void addVolumeToGroup(Long volumeId, String name) {
+    public void addVolumeToGroup(int volumeId, String name) {
 
     }
 
     @Override
-    public void removeVolumeFromGroup(Long volumeId, String name) {
+    public void removeVolumeFromGroup(int volumeId, String name) {
 
     }
